requirements.txt
aiofiles==0.4.0
aiohttp==3.5.4
asyncio==3.4.3
numpy==1.16.3
python-multipart==0.0.5
starlette==0.12.0
uvicorn==0.7.1
opencv-python==4.2.0.34
Pillow==7.1.2 
face-recognition==1.3.0
